Add nationality info to unknown ancestors
When we've gone as far back as we can on a family tree, we hit "unknown
ancestors". This commit allows these unknown ancestors to be assigned a
nationality. This will allow us to take into account that some ancestor
branches will not have a known nationality in our calculations.

To accomplish this, the generic binary tree has been converted into a
domain-specific `FamilyTree`.