Calculate children's nationality based on parents
Parents might not have the same total score since we don't always have a
full tree that goes as far back for every one. In order for both parents
to contribute equally, I scale up the one with the smaller score such
that its total score matches the one with the larger score.

All this math no longer works on just integers so I needed to change the
scores over to floats.