Update family tree when new nationality is chosen
When a new nationality is chosen for a node, set that all the terminal
ancestors to that nationality. This works because if you say that an
individual is 100% French, then it follows that all their ancestors must
also be 100% French. And since we calculate recursively from ancestors
to figure out the composite nationality of all descendents, we are safe
in only updating terminal ancestors.